2 new LabWorks done and some small changes to existing files

+++ LabWork№1/CurrencyConverter.py
@@ -1,5 +1,6 @@
-rubles = int(input())
+rubles = float(input())
+
 if rubles < 0:
-    print('invalid')
+    print("invalid")
 else:
     print(rubles * 68.33)
